SPlit clone and workspace override
