+++ src/test/java/harness/baseclasses/SingleWorkspaceUnitGcp.java
@@ -1,4 +1,4 @@
 package harness.baseclasses;
 
 /** Base class for GCP unit tests that only need a single workspace for all test methods. */
-public class SingleWorkspaceUnitGcp extends SingleWorkspaceUnit {} // TODO-Dex
+public class SingleWorkspaceUnitGcp extends SingleWorkspaceUnit {}
