+++ src/test/java/unit/Config.java
@@ -2,8 +2,10 @@ package unit;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.not;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assumptions.assumingThat;
 
 import bio.terra.cli.businessobject.Config.BrowserLaunchOption;
 import bio.terra.cli.businessobject.Config.CommandRunnerOption;
@@ -11,6 +13,7 @@ import bio.terra.cli.serialization.userfacing.UFLoggingConfig;
 import bio.terra.cli.serialization.userfacing.UFServer;
 import bio.terra.cli.serialization.userfacing.UFWorkspace;
 import bio.terra.cli.utils.Logger;
+import bio.terra.workspace.model.CloudPlatform;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import harness.TestCommand;
 import harness.TestCommand.Result;
@@ -30,6 +33,16 @@ import org.junit.jupiter.api.Test;
 /** Tests for the `terra config` commands. */
 @Tag("unit")
 public class Config extends SingleWorkspaceUnit {
+  private String getTableFormatValue(List<HashMap> result, String Option) {
+    var res =
+        result.stream()
+            .filter(x -> (x.get("option").equals(Option)))
+            .findFirst()
+            .orElse(new HashMap<>(Map.of("value", "")))
+            .get("value");
+    return res != null ? res.toString() : null;
+  }
+
   @Test
   @DisplayName("resource limit config throws error if exceeded")
   void resourceLimit() throws IOException {
@@ -300,13 +313,48 @@ public class Config extends SingleWorkspaceUnit {
         "workspace create affects config list in Json format");
   }
 
-  private String getTableFormatValue(List<HashMap> result, String Option) {
-    var res =
-        result.stream()
-            .filter(x -> (x.get("option").equals(Option)))
-            .findFirst()
-            .orElse(new HashMap<>(Map.of("value", "")))
-            .get("value");
-    return res != null ? res.toString() : null;
+  @Test
+  @DisplayName("app-launch config affects how apps are launched")
+  void appLaunchGcp() throws IOException {
+    assumingThat(
+        getCloudPlatform() == CloudPlatform.GCP,
+        () -> {
+          String badImageError = "No such image: badimageid";
+
+          workspaceCreator.login();
+
+          // `terra workspace set --id=$id`
+          TestCommand.runCommandExpectSuccess("workspace", "set", "--id=" + getUserFacingId());
+
+          // set the docker image id to an invalid string
+          TestCommand.runCommandExpectSuccess("config", "set", "image", "--image=badimageid");
+
+          // `terra config set app-launch LOCAL_PROCESS`
+          TestCommand.runCommandExpectSuccess("config", "set", "app-launch", "LOCAL_PROCESS");
+
+          // Apps launched via local process should not be affected
+          Result cmd = TestCommand.runCommand("app", "execute", "echo", "$GOOGLE_CLOUD_PROJECT");
+          // Cmd exit code can either be 0=success or 1 because gcloud fails with
+          // `(gcloud.config.get-value) Failed to create the default configuration. Ensure your have
+          // the
+          // correct permissions on`.
+          assertThat(
+              "Expected to return exit code 0 or 1, instead got " + cmd.exitCode,
+              cmd.exitCode == 0 || cmd.exitCode == 1);
+          assertThat(
+              "bad docker image should not affect local mode",
+              cmd.stdErr,
+              not(containsString(badImageError)));
+
+          // `terra config set app-launch DOCKER_CONTAINER`
+          TestCommand.runCommandExpectSuccess("config", "set", "app-launch", "DOCKER_CONTAINER");
+
+          // Apps launched via docker container should error out
+          String stdErr =
+              TestCommand.runCommandExpectExitCode(
+                  3, "app", "execute", "echo", "$GOOGLE_CLOUD_PROJECT");
+          assertThat(
+              "docker image not found error returned", stdErr, containsString(badImageError));
+        });
   }
 }
